Implement change requests for merge
