+++ NOTICE.md
@@ -1,6 +1,6 @@
 # Credits
 
-The asynchronous chunk loading executor code in `cubicchunks.server.chunkio.async` is based on parts of
+The asynchronous chunk loading executor code in `cubicchunks.server.chunkio.async.forge` is based on parts of
 [MinecraftForge](https://github.com/MinecraftForge/MinecraftForge) and licensed under the
 [GNU Lesser General Public License v2.1](http://www.gnu.org/licenses/lgpl-2.1.html):
   - [Original license text](https://github.com/MinecraftForge/MinecraftForge/blob/279380b4f1aa02a4df5e2b323c11ca6ab1531431/LICENSE-new.txt)
