+++ src/main/java/cubicchunks/asm/JvmNames.java
@@ -33,7 +33,6 @@ public class JvmNames {
 		BLOCK_POS = "Lnet/minecraft/util/math/BlockPos;",
 		CHUNK = "Lnet/minecraft/world/chunk/Chunk;",
 		CHUNK_CACHE = "Lnet/minecraft/world/ChunkCache;",
-		CHUNK_IO_EXECUTOR = "Lnet/minecraftforge/common/chunkio/ChunkIOExecutor;",
 		CLASS_INHERITANCE_MULTI_MAP = "Lnet/minecraft/util/ClassInheritanceMultiMap;",
 		COMMAND_BASE = "Lnet/minecraft/command/CommandBase;",
 		COMMAND_TELEPORT = "Lnet/minecraft/command/server/CommandTeleport;",
@@ -57,14 +56,12 @@ public class JvmNames {
 		BLOCK_POS_GETY = BLOCK_POS + "getY()I",
 		CHUNK_CACHE_GET_BLOCK_STATE = CHUNK_CACHE + "getBlockState(" + BLOCK_POS + ")" + IBLOCK_STATE,
 		CHUNK_GET_ENTITY_LISTS = CHUNK + "getEntityLists()[" + CLASS_INHERITANCE_MULTI_MAP,
-		CHUNK_IO_EXECUTOR_TICK = CHUNK_IO_EXECUTOR + "tick()V",
 		COMMAND_BASE_PARSE_DOUBLE = COMMAND_BASE + "parseDouble(D" + STRING + "IIZ)D",
 		COMMAND_TELEPORT_GET_ENTITY = COMMAND_TELEPORT + "getEntity(" + MINECRAFT_SERVER + ICOMMAND_SENDER + STRING + ")" + ENTITY,
 		COMMAND_TP_GET_COMMAND_SENDER_AS_PLAYER = COMMAND_TP + "getCommandSenderAsPlayer(" + ICOMMAND_SENDER + ")" + ENTITY_PLAYER_MP,
 		COMMAND_TP_GET_ENTITY = COMMAND_TP + "getEntity(" + MINECRAFT_SERVER + ICOMMAND_SENDER + STRING + ")" + ENTITY,
 		DEDICATED_SERVER_IS_ANNOUNCING_PLAYER_ACHIEVEMENTS = DEDICATED_SERVER + "isAnnouncingPlayerAchievements()Z",
 		GUI_OVERLAY_DEBUG_CALL = GUI_OVERLAY_DEBUG + "call()Ljava/util/List;",
-		MINECRAFT_SERVER_UPDATE_TIME_LIGHT_AND_ENTITIES = MINECRAFT_SERVER + "updateTimeLightAndEntities()V",
 		WORLD_CLIENT_GET_CHUNK_FROM_BLOCK_COORDS = WORLD_CLIENT + "getChunkFromBlockCoords(" + BLOCK_POS + ")" + CHUNK,
 		WORLD_GET_LIGHT_FOR = WORLD + "getLightFor(" + ENUM_SKY_BLOCK + BLOCK_POS + ")I",
 		WORLD_GET_LIGHT_WITH_FLAG = WORLD + "getLight(" + BLOCK_POS + "Z)I",
