--- src/main/java/cubicchunks/asm/mixin/core/server/MixinMinecraftServer_AsyncCubeIO.java
@@ -1,48 +1,0 @@
-/*
- *  This file is part of Cubic Chunks Mod, licensed under the MIT License (MIT).
- *
- *  Copyright (c) 2015 contributors
- *
- *  Permission is hereby granted, free of charge, to any person obtaining a copy
- *  of this software and associated documentation files (the "Software"), to deal
- *  in the Software without restriction, including without limitation the rights
- *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- *  copies of the Software, and to permit persons to whom the Software is
- *  furnished to do so, subject to the following conditions:
- *
- *  The above copyright notice and this permission notice shall be included in
- *  all copies or substantial portions of the Software.
- *
- *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- *  THE SOFTWARE.
- */
-
-package cubicchunks.asm.mixin.core.server;
-
-import cubicchunks.server.chunkio.async.AsyncWorldIOExecutor;
-import net.minecraft.server.MinecraftServer;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-import static cubicchunks.asm.JvmNames.CHUNK_IO_EXECUTOR_TICK;
-import static cubicchunks.asm.JvmNames.MINECRAFT_SERVER_UPDATE_TIME_LIGHT_AND_ENTITIES;
-
-@Mixin(MinecraftServer.class)
-public class MixinMinecraftServer_AsyncCubeIO {
-
-	/**
-	 * Add callback to tick the async cube loader. The server may host both vanilla and cubic worlds, so we tick our
-	 * chunk loader after the forge one
-	 */
-	@Inject(method=MINECRAFT_SERVER_UPDATE_TIME_LIGHT_AND_ENTITIES, at = @At(value = "INVOKE", target = CHUNK_IO_EXECUTOR_TICK, shift = At.Shift.AFTER))
-	private void tickCubeLoader(CallbackInfo info) {
-		AsyncWorldIOExecutor.tick();
-	}
-}
