+++ src/main/java/cubicchunks/server/chunkio/async/forge/AsyncIOProvider.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package cubicchunks.server.chunkio.async;
+package cubicchunks.server.chunkio.async.forge;
 
 import javax.annotation.Nullable;
 import java.util.concurrent.ConcurrentLinkedQueue;
