+++ src/main/java/cubicchunks/server/chunkio/async/forge/QueuedCube.java
@@ -17,32 +17,37 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package cubicchunks.server.chunkio.async;
+package cubicchunks.server.chunkio.async.forge;
 
 import com.google.common.base.Objects;
 import net.minecraft.world.World;
 
-class QueuedColumn {
+/**
+ * Taking from Sponge, with modifications
+ */
+class QueuedCube {
 	final int x;
+	final int y;
 	final int z;
 	final World world;
 
-	QueuedColumn(int x, int z, World world) {
+	QueuedCube(int x, int y, int z, World world) {
 		this.x = x;
+		this.y = y;
 		this.z = z;
 		this.world = world;
 	}
 
 	@Override
 	public int hashCode() {
-		return (x * 31 + z * 29) ^ world.hashCode();
+		return (x * 31 + y * 23 * z * 29) ^ world.hashCode();
 	}
 
 	@Override
 	public boolean equals(Object object) {
-		if (object instanceof QueuedColumn) {
-			QueuedColumn other = (QueuedColumn) object;
-			return x == other.x && z == other.z && world == other.world;
+		if (object instanceof QueuedCube) {
+			QueuedCube other = (QueuedCube) object;
+			return x == other.x && y == other.y && z == other.z && world == other.world;
 		}
 
 		return false;
@@ -53,6 +58,7 @@ class QueuedColumn {
 		return Objects.toStringHelper(this)
 				.addValue(this.world)
 				.add("x", this.x)
+				.add("y", this.y)
 				.add("z", this.z)
 				.toString();
 	}
